+++ internal/coordination/client.go
@@ -369,11 +369,11 @@ func (c *Client) closeSessions(ctx context.Context) {
 func defaultCreateSessionConfig() *options.CreateSessionOptions {
 	return &options.CreateSessionOptions{
 		Description:             "YDB Go SDK",
-		SessionTimeout:          time.Second * 5, //nolint:gomnd
+		SessionTimeout:          time.Second * 5, //nolint:mnd
 		SessionStartTimeout:     time.Second * 1,
 		SessionStopTimeout:      time.Second * 1,
-		SessionKeepAliveTimeout: time.Second * 10,       //nolint:gomnd
-		SessionReconnectDelay:   time.Millisecond * 500, //nolint:gomnd
+		SessionKeepAliveTimeout: time.Second * 10,       //nolint:mnd
+		SessionReconnectDelay:   time.Millisecond * 500, //nolint:mnd
 	}
 }