+++ internal/discovery/discovery.go
@@ -57,7 +57,7 @@ func Discover(
 
 	if response.GetOperation().GetStatus() != Ydb.StatusIds_SUCCESS {
 		return nil, location, xerrors.WithStackTrace(
-			xerrors.FromOperation(response.GetOperation()),
+			xerrors.Operation(xerrors.FromOperation(response.GetOperation())),
 		)
 	}
 
@@ -151,9 +151,9 @@ func (c *Client) WhoAmI(ctx context.Context) (whoAmI *discovery.WhoAmI, err erro
 
 	if response.GetOperation().GetStatus() != Ydb.StatusIds_SUCCESS {
 		return nil, xerrors.WithStackTrace(
-			xerrors.FromOperation(
+			xerrors.Operation(xerrors.FromOperation(
 				response.GetOperation(),
-			),
+			)),
 		)
 	}