+++ internal/table/scanner/scanner.go
@@ -1235,7 +1235,7 @@ func (s *valueScanner) errorf(depth int, f string, args ...interface{}) error {
 
 func (s *valueScanner) typeError(act, exp interface{}) {
 	_ = s.errorf(
-		2, //nolint:gomnd
+		2, //nolint:mnd
 		"unexpected types during scan at %q %s: %s; want %s",
 		s.path(),
 		s.getType(),
@@ -1247,7 +1247,7 @@ func (s *valueScanner) typeError(act, exp interface{}) {
 func (s *valueScanner) valueTypeError(act, exp interface{}) {
 	// unexpected value during scan at \"migration_status\" Int64: NullFlag; want Int64
 	_ = s.errorf(
-		2, //nolint:gomnd
+		2, //nolint:mnd
 		"unexpected value during scan at %q %s: %s; want %s",
 		s.path(),
 		s.getType(),
@@ -1258,7 +1258,7 @@ func (s *valueScanner) valueTypeError(act, exp interface{}) {
 
 func (s *valueScanner) notFoundColumnByIndex(idx int) error {
 	return s.errorf(
-		2, //nolint:gomnd
+		2, //nolint:mnd
 		"not found %d column",
 		idx,
 	)
@@ -1266,7 +1266,7 @@ func (s *valueScanner) notFoundColumnByIndex(idx int) error {
 
 func (s *valueScanner) notFoundColumnName(name string) error {
 	return s.errorf(
-		2, //nolint:gomnd
+		2, //nolint:mnd
 		"not found column '%s'",
 		name,
 	)
@@ -1274,7 +1274,7 @@ func (s *valueScanner) notFoundColumnName(name string) error {
 
 func (s *valueScanner) noColumnError(name string) error {
 	return s.errorf(
-		2, //nolint:gomnd
+		2, //nolint:mnd
 		"no column %q",
 		name,
 	)
@@ -1282,7 +1282,7 @@ func (s *valueScanner) noColumnError(name string) error {
 
 func (s *valueScanner) overflowError(i, n interface{}) error {
 	return s.errorf(
-		2, //nolint:gomnd
+		2, //nolint:mnd
 		"overflow error: %d overflows capacity of %t",
 		i,
 		n,