+++ topic/topicwriter/topicwriter.go
@@ -14,8 +14,8 @@ var (
 	// Deprecated: the error will not be returned. Topic writer allow overflow queue for single call.
 	// Will be removed after October 2025.
 	// Read about versioning policy: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#deprecated
-	ErrQueueLimitExceed                      = topicwriterinternal.PublicErrQueueIsFull
-	ErrMessagesPutToInternalQueueBeforeError = topicwriterinternal.PublicErrMessagesPutToInternalQueueBeforeError
+	ErrQueueLimitExceed                      = topicwriterinternal.ErrPublicQueueIsFull
+	ErrMessagesPutToInternalQueueBeforeError = topicwriterinternal.ErrPublicMessagesPutToInternalQueueBeforeError
 )
 
 // Writer represent write session to topic